Fix used_segspace counter leak
GPDB currently has 2 APIs of creating work files

1. Using workfile_mgr :

  1. This uses a workfile set to contain all the related workfiles.
  2. All workfiles are created under the same directory. As a result,
     when it is time to delete the workfiles, we can simply delete the directory.
  3. There is no mechanism to control the exact name of the workfile created.
     All workfiles are referenced using an ID. `"%s/spillfile\_f%u", work\_set-\>path, file\_no);`
  4. Deletion of workiles is automatically handled in case of an exception,
     as all the files are contained in one directory, thus no counter/files are leaked.

2. Using ExecWorkfile :

  1. There is a Create/Open/Read/Write/Flush/Close API.
  2. This is the lower-level API also used by workfile_mgr to manage the workfiles directly.
  3. Each operator is resposible to correctly close the workfile it creates.

When we write to a workfile, we reserve and add the number of bytes to used_segspace,
and add that amount to `workfile->size`. Then, we call `workfile_update_in_progress_size`.
If the workfile has an owning workset, we update the workset's `in_progress_size` counter.
However, if it has no such workset, we completely ignore the count -
and this is what ultimately leaks the `used_segspace` counter.

There are few places in GPDB that use the lower level API to spill:
1. `LogicalTapeSet`
2. `Tuplesortstate`
3. `ntuplestore_create_spill_files`
4. `AppendOnlyVisimapDelete`
5. `tuplesort_begin_heap_file_readerwriter_mk`

In normal execution, this does not leak because `ExecutorEnd` is called correctly -
and that deletes the files, and updates the counters correctly. However when an exception
occurs during execution, `ExecutorEnd` is only called for `SELECT` queries.
For `UPDATE`, `INSERT`, `DELETE` queries- we don't call `ExecutorEnd` and thus leak files and the counter.

This is a quick fix to account for bytes written in workfiles that are not part of any workfile set -
that is those created using the lower level API. Hence when we abort or error out,
we subtract this amount from the `used_segspace` counter.

In the long term, having two APIs is cumbersome and error-prone, and we have created a story to fix it.

As part of this fix, we also add a failure injection for `workfile_write_failure`.
As well as test setup and clean up for segspace to change config and restart gpdb

New scenarios covering `used_segspace` counter leak are added:
- Case 1: UPDATE with Sort (non-mk)
- Case 2: UPDATE with SHARED_SORT_XSLICE
- Case 4: UPDATE with CTE and mk_sort
- Case 5: DELETE with append-only tables

We cannot repro the leak in Case 3, due to Materialize properly release the counter even under transaction
abort through different code path:

130 1    0x102b97e65 postgres errstart + 0x5d5
131 2    0x102c3a62b postgres adjust_size_temp_file_new + 0x1eb
132 3    0x102c3a3ed postgres update_workset_size + 0x3d
133 4    0x102c3a36e postgres workfile_mgr_close_file + 0x2be
134 5    0x102c0b5e6 postgres ntuplestore_cleanup + 0x156
135 6    0x102c0a75f postgres XCallBack_NTS + 0x1f
136 7    0x10251d529 postgres CallXactCallbacksOnce + 0x99
137 8    0x10251da95 postgres AbortTransaction + 0x3e5
138 9    0x10251fe77 postgres AbortCurrentTransaction + 0x87
139 10   0x102a064a9 postgres PostgresMain + 0x939
140 11   0x10297d1dd postgres BackendRun + 0x2dd
141 12   0x10297c61a postgres BackendStartup + 0x17a
142 13   0x102977d1f postgres ServerLoop + 0x51f
143 14   0x102975fc7 postgres PostmasterMain + 0x1427
144 15   0x1028532d4 postgres main + 0x334
145 16   0x7fff8992b5c9 libdyld.dylib start + 0x1

[#130306049]

Signed-off-by: Xin Zhang <xzhang@pivotal.io>