Make sure CallXactCallbacksOnce are executed during preparetransaction as well.
All of the registered callbacks are just for cleaning up temp files. So it
is ok to delete those files during PrepareTransaction.
The reason we do this change now is because BFZ relies on this callback to
clean up its data structure. This BFZ data structure uses
DataSource context for its memory. DataSource Context is
allocated on TopTransactionMemoryContext. By making this change, we made
sure that we delete the data structure before
TopTransactionMemoryContext / DataSource context is deleted.