Refactor gfile and bfz to use `File` instead of file descriptor.
This refactors gfile to use `File` so file deletion will happen through
ResourceOwner. This is based on the commit 6c5eba7ca87e4020d8cb643cfa8f98c2382a032b.
With this refactor, we can leverage gile's zlib API for BFZ instead of
having separate code in compress_zlib.c

Note: There are still possible memory leak as some of the data
structures are allocated in TopMemoryContext. This needs to be fixed in
seperate commit.