Abstract file access api from gfile.
This will enable gfile to work with low-level file descriptor for gpfdist and File
for workfile to handle resource cleanup properly.
Note: When transaction abort, it could have possible memory leak as some of the data
structures(bfz) are allocated in TopMemoryContext and this commit remove
bfz_close_callback. This needs to be fixed in
seperate commit.